feat: add dashboard compliance endpoint and unify placeholder audit
