+++ docs/DATABASE_FIRST_USAGE_GUIDE.md
@@ -65,7 +65,7 @@ The Flask dashboard exposes a `/dashboard/compliance` endpoint that reads these
 metrics and shows real-time placeholder removal progress. When a placeholder is corrected, record the update in `analytics.db:correction_logs`. This ensures future audits can cross-reference removed placeholders with generated fixes.
 
 ### Placeholder Correction Workflow
-1. Run `scripts/audit_codebase_placeholders.py` to log all TODOs.
+1. Run `scripts/code_placeholder_audit.py` to log all TODOs.
 2. Review entries in `analytics.db:placeholder_audit` and fix the code.
 3. Record corrections with `scripts/correction_logger_and_rollback.py` for audit.
 4. Monitor `/dashboard/compliance` to verify the compliance score improves.
