+++ scripts/audit_codebase_placeholders.py
@@ -1,310 +1,12 @@
-"""
-Enterprise Database-First Code Audit Script
-
-MANDATORY REQUIREMENTS:
-1. Query production.db for tracked file patterns and audit templates.
-2. Traverse all files, extract line number, context, and type.
-3. Log each finding to analytics.db.todo_fixme_tracking with file path, line number, context, timestamp.
-4. Update /dashboard/compliance with removal status and compliance metrics.
-5. Include tqdm progress bar, start time logging, timeout, ETC calculation, anti-recursion validation.
-6. DUAL COPILOT: Secondary validator checks audit completeness and compliance.
-7. Integrate best practices from deep web research for codebase audits.
-"""
-
+"""Deprecated wrapper for :mod:`scripts.code_placeholder_audit`."""
 from __future__ import annotations
 
-import json
-import logging
-import os
-import re
-import sqlite3
-import time
-from datetime import datetime
-from pathlib import Path
-from typing import Dict, List, Optional
+from scripts.code_placeholder_audit import main
 
-from tqdm import tqdm
-
-from scripts.continuous_operation_orchestrator import (
-    validate_enterprise_operation,
-)
-from utils.log_utils import log_message
-from scripts.database.add_code_audit_log import ensure_code_audit_log
-
-# Visual processing indicator constants
-TEXT = {
-    "start": "[START]",
-    "success": "[SUCCESS]",
-    "error": "[ERROR]",
-    "info": "[INFO]",
-    "progress": "[PROGRESS]",
-    "validate": "[VALIDATE]",
-    "complete": "[COMPLETE]",
-}
-
-DEFAULT_PATTERNS = [
-    r"TODO",
-    r"FIXME",
-]
-
-
-# Database-first: fetch placeholder patterns from production.db
-def fetch_db_placeholders(production_db: Path) -> List[str]:
-    """Fetch placeholder patterns from production.db for audit."""
-    if not production_db.exists():
-        return []
-    with sqlite3.connect(production_db) as conn:
-        try:
-            cur = conn.execute("SELECT placeholder_name FROM template_placeholders")
-            return [row[0] for row in cur.fetchall()]
-        except Exception as e:
-            log_message(
-                f"{TEXT['error']} Failed to fetch placeholders from production.db: {e}",
-                level=logging.ERROR,
-            )
-            return []
-
-
-# Insert findings into analytics.db.code_audit_log
-def log_findings(results: List[Dict], analytics_db: Path, simulate: bool = False) -> None:
-    """Log audit results to analytics.db.
-
-    Parameters
-    ----------
-    results : List[Dict]
-        Findings collected from the codebase scan.
-    analytics_db : Path
-        Target analytics database path.
-    simulate : bool, optional
-        If True, no database writes occur.
-
-    Returns
-    -------
-    None
-        The function writes to the database when not in simulation mode.
-    """
-    analytics_db.parent.mkdir(parents=True, exist_ok=True)
-    ensure_code_audit_log(analytics_db)
-    if simulate:
-        log_message("[TEST MODE] Simulation enabled: not writing to analytics.db")
-        return
-    with sqlite3.connect(analytics_db) as conn:
-        conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS code_audit_log (
-                id INTEGER PRIMARY KEY,
-                file_path TEXT NOT NULL,
-                line_number INTEGER,
-                placeholder_type TEXT,
-                context TEXT,
-                timestamp TEXT NOT NULL
-            )
-            """
-        )
-        conn.execute("CREATE INDEX IF NOT EXISTS idx_code_audit_log_file_path ON code_audit_log(file_path)")
-        conn.execute("CREATE INDEX IF NOT EXISTS idx_code_audit_log_timestamp ON code_audit_log(timestamp)")
-        conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS todo_fixme_tracking (
-                file_path TEXT,
-                line_number INTEGER,
-                placeholder_type TEXT,
-                context TEXT,
-                timestamp DATETIME DEFAULT CURRENT_TIMESTAMP
-            )
-            """
-        )
-        for row in results:
-            values = (
-                row["file"],
-                row["line"],
-                row["pattern"],
-                row["context"],
-                datetime.now().isoformat(),
-            )
-            conn.execute(
-                "INSERT INTO code_audit_log (file_path, line_number, placeholder_type, context, timestamp)"
-                " VALUES (?, ?, ?, ?, ?)",
-                values,
-            )
-            conn.execute(
-                "INSERT INTO todo_fixme_tracking (file_path, line_number, placeholder_type, context, timestamp)"
-                " VALUES (?, ?, ?, ?, ?)",
-                (
-                    row["file"],
-                    row["line"],
-                    row["pattern"],
-                    row["context"],
-                    datetime.now().isoformat(),
-                ),
-            )
-        conn.commit()
-
-
-# Update dashboard/compliance with summary JSON
-def update_dashboard(count: int, dashboard_dir: Path) -> None:
-    """Write summary JSON to dashboard/compliance directory."""
-    dashboard_dir.mkdir(parents=True, exist_ok=True)
-    compliance = max(0, 100 - count)
-    status = "complete" if count == 0 else "issues_pending"
-    data = {
-        "timestamp": datetime.now().isoformat(),
-        "findings": count,
-        "compliance_score": compliance,
-        "progress_status": status,
-    }
-    summary_file = dashboard_dir / "placeholder_summary.json"
-    summary_file.write_text(json.dumps(data, indent=2), encoding="utf-8")
-
-
-# Scan files for patterns with timeout and visual indicators
-def scan_files(workspace: Path, patterns: List[str], timeout: Optional[float] = None) -> List[Dict]:
-    """Scan files for given patterns with optional timeout and progress bar."""
-    results: List[Dict] = []
-    files = [f for f in workspace.rglob("*") if f.is_file()]
-    start = time.time()
-    with tqdm(total=len(files), desc=f"{TEXT['progress']} scanning", unit="file") as bar:
-        for file in files:
-            if timeout and time.time() - start > timeout:
-                bar.write(f"{TEXT['error']} timeout reached")
-                break
-            try:
-                lines = file.read_text(encoding="utf-8", errors="ignore").splitlines()
-            except Exception as e:
-                log_message(
-                    f"{TEXT['error']} Could not read {file}: {e}",
-                    level=logging.ERROR,
-                )
-                bar.update(1)
-                continue
-            for idx, line in enumerate(lines, 1):
-                for pat in patterns:
-                    if re.search(pat, line):
-                        results.append(
-                            {
-                                "file": str(file),
-                                "line": idx,
-                                "pattern": pat,
-                                "context": line.strip()[:200],
-                            }
-                        )
-            bar.update(1)
-    return results
-
-
-# DUAL COPILOT: Secondary validator for audit completeness
-def validate_results(expected_count: int, analytics_db: Path) -> bool:
-    """Secondary validation step for dual copilot pattern."""
-    with sqlite3.connect(analytics_db) as conn:
-        cur = conn.execute("SELECT COUNT(*) FROM todo_fixme_tracking")
-        db_count = cur.fetchone()[0]
-    return db_count >= expected_count
-
-
-def calculate_etc(start_time: float, current_progress: int, total_work: int) -> str:
-    """Calculate estimated time to completion."""
-    elapsed = time.time() - start_time
-    if current_progress > 0:
-        total_estimated = elapsed / (current_progress / total_work)
-        remaining = total_estimated - elapsed
-        return f"{remaining:.2f}s remaining"
-    return "N/A"
-
-
-def main(
-    workspace_path: Optional[str] = None,
-    analytics_db: Optional[str] = None,
-    production_db: Optional[str] = None,
-    dashboard_dir: Optional[str] = None,
-    timeout_minutes: int = 30,
-    simulate: bool = False,
-) -> bool:
-    """Entry point for placeholder auditing with full enterprise compliance.
-
-    Parameters
-    ----------
-    simulate:
-        If ``True``, skip writing to databases and dashboard.
-    """
-    # Visual processing indicators: start time, process ID, anti-recursion validation
-    start_time = time.time()
-    process_id = os.getpid()
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-    log_message(f"{TEXT['start']} auditing workspace")
-    log_message(f"Start Time: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
-    log_message(f"Process ID: {process_id}")
-
-    # Anti-recursion validation
-    if os.getenv("GH_COPILOT_DISABLE_VALIDATION") != "1":
-        validate_enterprise_operation()
-
-    workspace = Path(workspace_path or os.getenv("GH_COPILOT_WORKSPACE", Path.cwd()))
-    analytics = Path(analytics_db or workspace / "databases" / "analytics.db")
-    production = Path(production_db or workspace / "databases" / "production.db")
-    dashboard = Path(dashboard_dir or workspace / "dashboard" / "compliance")
-
-    # Database-first: fetch patterns from production.db
-    patterns = DEFAULT_PATTERNS + fetch_db_placeholders(production)
-    timeout = timeout_minutes * 60 if timeout_minutes else None
-
-    # Scan files with progress bar and ETC calculation
-    files = [f for f in workspace.rglob("*") if f.is_file()]
-    results: List[Dict] = []
-    with tqdm(total=len(files), desc=f"{TEXT['progress']} scanning", unit="file") as bar:
-        for idx, file in enumerate(files, 1):
-            if timeout and time.time() - start_time > timeout:
-                bar.write(f"{TEXT['error']} timeout reached")
-                break
-            try:
-                lines = file.read_text(encoding="utf-8", errors="ignore").splitlines()
-            except Exception as e:
-                log_message(
-                    f"{TEXT['error']} Could not read {file}: {e}",
-                    level=logging.ERROR,
-                )
-                bar.update(1)
-                continue
-            for line_num, line in enumerate(lines, 1):
-                for pat in patterns:
-                    if re.search(pat, line):
-                        results.append(
-                            {
-                                "file": str(file),
-                                "line": line_num,
-                                "pattern": pat,
-                                "context": line.strip()[:200],
-                            }
-                        )
-            elapsed = time.time() - start_time
-            etc = calculate_etc(start_time, idx, len(files))
-            bar.set_postfix(ETC=etc)
-            bar.update(1)
-
-    # Log findings to analytics.db
-    log_findings(results, analytics, simulate=simulate)
-    # Update dashboard/compliance
-    if not simulate:
-        update_dashboard(len(results), dashboard)
-    else:
-        log_message("[TEST MODE] Dashboard update skipped")
-    elapsed = time.time() - start_time
-    log_message(f"{TEXT['info']} audit completed in {elapsed:.2f}s")
-
-    # DUAL COPILOT validation
-    valid = validate_results(len(results), analytics)
-    if valid:
-        log_message(f"{TEXT['success']} audit logged {len(results)} findings")
-    else:
-        log_message(
-            f"{TEXT['error']} validation mismatch",
-            level=logging.ERROR,
-        )
-    return valid
-
+__all__ = ["main"]
 
 if __name__ == "__main__":
     import argparse
-
     parser = argparse.ArgumentParser(description="Audit workspace for TODO/FIXME placeholders")
     parser.add_argument("--workspace-path", type=str, help="Workspace to scan")
     parser.add_argument("--analytics-db", type=str, help="analytics.db location")
@@ -313,6 +15,7 @@ if __name__ == "__main__":
     parser.add_argument("--timeout-minutes", type=int, default=30, help="Scan timeout in minutes")
     parser.add_argument("--simulate", action="store_true", help="Run in test mode without writes")
     args = parser.parse_args()
+
     success = main(
         workspace_path=args.workspace_path,
         analytics_db=args.analytics_db,
